+++ src/Console/Application.php
@@ -111,6 +111,7 @@ class Application extends BaseApplication
         $commands[] = new \SugarCli\Console\Command\Relationships\RelationshipsLoadCommand();
         $commands[] = new \SugarCli\Console\Command\Relationships\RelationshipsStatusCommand();
         $commands[] = new \SugarCli\Console\Command\SystemQuickRepairCommand();
+        $commands[] = new \SugarCli\Console\Command\SelfUpdateCommand();
         $commands[] = new \SugarCli\Console\Command\User\ListCommand();
         $commands[] = new \SugarCli\Console\Command\User\UpdateCommand();
         return $commands;